requirements.txt
Django>=5.0,<6.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
model-bakery>=1.18.0
django-debug-toolbar>=4.4.0
django-extensions>=3.2.3
coverage>=7.4.0